requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

streamlit
pandas
numpy==1.26.4
requests
faiss-cpu==1.7.4
accelerate==0.27.2
rank-bm25
python-dateutil
sentence-transformers==2.6.1
transformers==4.37.2
torch==2.1.2+cpu
groq
python-dotenv
rapidfuzz
openpyxl
reportlab
PyPDF2
python-docx
xlsxwriter
beautifulsoup4
openai>=1.40.0
httpx==0.27.0
pymssql



